@@ -3,7 +3,7 @@ description: A new Flutter project.
 publish_to: none
 version: 1.0.0+1
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
   flutter:
     sdk: flutter
@@ -31,12 +31,12 @@ dependencies:
   # pedantic: ^1.9.2
   provider: ^5.0.0
   shimmer: ^2.0.0-nullsafety.0
-  
+  build_runner: ^1.10.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-build_runner: ^1.10.0   
+   
 flutter:
   uses-material-design: true
   assets:
