@@ -4,7 +4,7 @@ import '../../init/cache/locale_manager.dart';
 import '../../init/navigation/navigation_service.dart';
 
 abstract class BaseViewModel {
-  BuildContext context;
+  late BuildContext context;
 
   LocaleManager localeManager = LocaleManager.instance;
   NavigationService navigation = NavigationService.instance;
