@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
 class LanguageManager {
-  static LanguageManager _instance;
+  static LanguageManager? _instance;
   static LanguageManager get instance {
     if (_instance == null) _instance = LanguageManager._init();
-    return _instance;
+    return _instance!;
   }
 
   LanguageManager._init();
