@@ -4,10 +4,10 @@ import 'package:provider/single_child_widget.dart';
 import 'theme_notifier.dart';
 
 class ApplicationProvider {
-  static ApplicationProvider _instance;
+  static ApplicationProvider? _instance;
   static ApplicationProvider get instance {
     if (_instance == null) _instance = ApplicationProvider._init();
-    return _instance;
+    return _instance!;
   }
 
   ApplicationProvider._init();
