@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class TextThemeLight {
-  static TextThemeLight _instace;
-  static TextThemeLight get instance {
+  static TextThemeLight? _instace;
+  static TextThemeLight? get instance {
     if (_instace == null) _instace = TextThemeLight._init();
     return _instace;
   }
