@@ -1,7 +1,7 @@
-import 'package:flutter_template/view/map/model/map_place_model.dart';
+import '../../map/model/map_place_model.dart';
 
 class Alarms {
-  int alarmId;
-  bool isAlarmActive;
-  MapPlace mapPlace;
+  int? alarmId;
+  bool? isAlarmActive;
+  MapPlace? mapPlace;
 }
