@@ -1,4 +1,4 @@
 abstract class INavigationService {
-  Future<void> navigateToPage({String path, Object data});
-  Future<void> navigateToPageClear({String path, Object data});
+  Future<void> navigateToPage({String? path, Object? data});
+  Future<void> navigateToPageClear({String? path, Object? data});
 }
