@@ -12,13 +12,13 @@ class NavigationService implements INavigationService {
   GlobalKey<NavigatorState> navigatorKey = GlobalKey();
 
   @override
-  Future<void> navigateToPage({String path, Object data}) async {
-    await navigatorKey.currentState.pushNamed(path, arguments: data);
+  Future<void> navigateToPage({String? path, Object? data}) async {
+    await navigatorKey.currentState!.pushNamed(path!, arguments: data);
   }
 
   @override
-  Future<void> navigateToPageClear({String path, Object data}) async {
-    await navigatorKey.currentState
-        .pushNamedAndRemoveUntil(path, removeAllOldRoutes, arguments: data);
+  Future<void> navigateToPageClear({String? path, Object? data}) async {
+    await navigatorKey.currentState!
+        .pushNamedAndRemoveUntil(path!, removeAllOldRoutes, arguments: data);
   }
 }
